git-svn-id: svn://123.57.173.36/meijia-parent@1594 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-service/src/main/java/com/simi/service/impl/user/UserActionRecordServiceImpl.java
@@ -22,12 +22,12 @@ public class UserActionRecordServiceImpl implements UserActionRecordService {
 	}
 
 	@Override
-	public int insert(UserActionRecord record) {
+	public Long insert(UserActionRecord record) {
 		return userActionRecordMapper.insert(record);
 	}
 
 	@Override
-	public int insertSelective(UserActionRecord record) {
+	public Long insertSelective(UserActionRecord record) {
 		return userActionRecordMapper.insertSelective(record);
 	}
